requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0